Combine if-statements and reject promise on false

+++ AngularJSAuthentication.Web/app/services/authService.js
@@ -83,26 +83,25 @@ app.factory('authService', ['$q', '$injector', 'localStorageService', 'ngAuthSet
        
         var authData = localStorageService.get('authorizationData');
 
-        if (authData) {
+        if (authData && authData.useRefreshTokens) {
 
-            if (authData.useRefreshTokens) {
+            var data = "grant_type=refresh_token&refresh_token=" + authData.refreshToken + "&client_id=" + ngAuthSettings.clientId;
 
-                var data = "grant_type=refresh_token&refresh_token=" + authData.refreshToken + "&client_id=" + ngAuthSettings.clientId;
+            localStorageService.remove('authorizationData');
 
-                localStorageService.remove('authorizationData');
+            $http = $http || $injector.get('$http');
+            $http.post(serviceBase + 'token', data, { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } }).success(function (response) {
 
-                $http = $http || $injector.get('$http');
-                $http.post(serviceBase + 'token', data, { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } }).success(function (response) {
+                localStorageService.set('authorizationData', { token: response.access_token, userName: response.userName, refreshToken: response.refresh_token, useRefreshTokens: true });
 
-                    localStorageService.set('authorizationData', { token: response.access_token, userName: response.userName, refreshToken: response.refresh_token, useRefreshTokens: true });
+                deferred.resolve(response);
 
-                    deferred.resolve(response);
-
-                }).error(function (err, status) {
-                    _logOut();
-                    deferred.reject(err);
-                });
-            }
+            }).error(function (err, status) {
+                _logOut();
+                deferred.reject(err);
+            });
+        } else {
+            deferred.reject();
         }
 
         return deferred.promise;
